requirements.txt
deepgram-sdk==3.2.7
python-dotenv
